# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: May Chicken & Burger Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu/items:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List menu items
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: available
          in: query
          schema:
            type: boolean
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
  /menu/items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a single menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found

  # ── PUBLIC RESTAURANT INFO ────────────────────────────────────────────────
...
  /restaurant/orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new customer order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
  /restaurant/orders/{id}:
    get:
      operationId: getOrderStatus
      tags: [orders]
      summary: Track order status by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /restaurant/coupons/validate:
    post:
      operationId: validateCoupon
      tags: [orders]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidationInput"
      responses:
        "200":
          description: Coupon valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "404":
          description: Invalid or expired coupon

  # ── AUTH ─────────────────────────────────────────────────────────────────
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
...
  /kitchen/orders/{id}/status:
    patch:
      operationId: updateKitchenOrderStatus
      tags: [kitchen]
      summary: Update order status from kitchen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KitchenOrderStatusPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug, sortOrder, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        itemCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    MenuItem:
      type: object
      required: [id, name, price, categoryId, available, featured, sortOrder, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        categoryId:
          type: integer
        available:
          type: boolean
        featured:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        category:
          $ref: "#/components/schemas/Category"
        createdAt:
          type: string
          format: date-time
...
    Coupon:
      type: object
      required: [id, code, discountType, discountValue, active, usageCount, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        description:
          type: ["string", "null"]
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minOrder:
          type: number
        active:
          type: boolean
        expiresAt:
          type: ["string", "null"]
          format: date-time
        usageCount:
          type: integer
        maxUsage:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    CouponValidationInput:
      type: object
      required: [code, orderTotal]
      properties:
        code:
          type: string
        orderTotal:
          type: number

    # ── ORDERS ────────────────────────────────────────────────────────────
    OrderItemInput:
      type: object
      required: [menuItemId, quantity]
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
          minimum: 1
    OrderItem:
      type: object
      required: [id, itemName, itemPrice, quantity, lineTotal]
      properties:
        id:
          type: integer
        menuItemId:
          type: ["integer", "null"]
        itemName:
          type: string
        itemPrice:
          type: number
        quantity:
          type: integer
        lineTotal:
          type: number
    OrderInput:
      type: object
      required: [orderType, customerName, customerPhone, items]
      properties:
        orderType:
          type: string
          enum: [delivery, pickup]
        customerName:
          type: string
          minLength: 1
        customerPhone:
          type: string
          minLength: 1
        customerEmail:
          type: ["string", "null"]
        deliveryAddress:
          type: string
        postalCode:
          type: string
        city:
          type: string
        notes:
          type: string
        couponCode:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
          minItems: 1
    Order:
      type: object
      required: [id, orderNumber, orderType, status, customerName, customerPhone, subtotal, deliveryFee, discountAmount, total, items, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        orderType:
          type: string
          enum: [delivery, pickup]
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivering, completed, cancelled]
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        deliveryAddress:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        subtotal:
          type: number
        deliveryFee:
          type: number
        discountAmount:
          type: number
        total:
          type: number
        couponCode:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
...
    KitchenOrderStatusPatch:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [confirmed, preparing, ready, delivering, completed]

    # ── CUSTOMERS ─────────────────────────────────────────────────────────
...
    AdminCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean
        username:
          type: ["string", "null"]

    # ── SETTINGS ─────────────────────────────────────────────────────────
...