# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * May Chicken & Burger Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderOrderType } from './orderOrderType';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  orderNumber: string;
  orderType: OrderOrderType;
  status: OrderStatus;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  deliveryAddress?: string | null;
  /** @nullable */
  postalCode?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  notes?: string | null;
  subtotal: number;
  deliveryFee: number;
  discountAmount: number;
  total: number;
  /** @nullable */
  couponCode?: string | null;
  items: OrderItem[];
  createdAt: Date;
}